# code/prices-domain/src/main/java/com/inditex/prices/domain/model/Price.java
package com.inditex.prices.domain.model;

import java.math.BigDecimal;
import java.time.LocalDateTime;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class Price {

    private Long productId;

    private Long brandId;

    private Long priceList;

    private LocalDateTime startDate;

    private LocalDateTime endDate;

    private BigDecimal price;

    private Integer priority;

    private String currency;

}